requirements.txt
Flask
SpeechRecognition
google-generativeai
python-dotenv
